+++ app/shared/src/commonMain/kotlin/com/example/merchapp/Database.kt
@@ -11,15 +11,15 @@ import kotlinx.serialization.Serializable
 
 class Merchant: RealmObject {
     @PrimaryKey
-    var id: Int = 0
+    var id: Long = 0
     var name: String = ""
 }
 
 class Purchaseable : RealmObject {
     @PrimaryKey
-    var id: Int = 0
-    var merchantId: Int = 0
-    var productId: Int = 0
+    var id: Long = 0
+    var merchantId: Long = 0
+    var productId: Long = 0
     var name: String = ""
     var productName: String = ""
     var thumbnail: String = ""
@@ -27,17 +27,17 @@ class Purchaseable : RealmObject {
 
 class PurchaseableView: RealmObject {
     @PrimaryKey
-    var id: Int = 0
-    var purchaseableId: Int = 0
+    var id: Long = 0
+    var purchaseableId: Long = 0
     var designName: String = ""
-    var background: Int = 0
+    var background: Long = 0
     var thumbnail: String = ""
     var name: String = ""
 }
 
 class Cart: RealmObject {
     @PrimaryKey
-    var id: Int = 0
+    var id: Long = 0
 }
 
 class Purchase : RealmObject {
@@ -65,7 +65,7 @@ object Database {
         }
     }
 
-    fun merchant(id: Int): Merchant? {
+    fun merchant(id: Long): Merchant? {
         return realm.query<Merchant>("id = $id").first().find()
     }
 
@@ -81,15 +81,15 @@ object Database {
         }
     }
 
-    fun purchaseables(merchantId: Int): Array<Purchaseable> {
+    fun purchaseables(merchantId: Long): Array<Purchaseable> {
         return realm.query<Purchaseable>("merchantId = $0", merchantId).find().toTypedArray()
     }
 
-    fun purchaseable(id: Int): Purchaseable? {
+    fun purchaseable(id: Long): Purchaseable? {
         return realm.query<Purchaseable>("id = $0", id).first().find()
     }
 
-    fun views(purchaseableId: Int): Array<PurchaseableView> {
+    fun views(purchaseableId: Long): Array<PurchaseableView> {
         return realm.query<PurchaseableView>("purchaseableId = $0", purchaseableId).find().toTypedArray()
     }
